data/requirements.txt
SQLAlchemy
asyncmy
aiomysql
Faker
cryptography
